Peer reconnect support.
TODO After reconnect the clients after first connection are left without peer server side messages.
