+++ mles/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mles"
-version = "0.7.3"
+version = "0.7.4"
 authors = ["jq-rs"]
 description = "Mles - a distributed publish/subscribe database server"
 license = "GPL-3.0"
